Move questionnaire entity, remove condition entity
