+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/config/WebSecurityConfig.java
@@ -21,8 +21,6 @@ public class WebSecurityConfig {
                 .and()
                 .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 .and()
-                .csrf()
-                .disable()
                 .build();
     }
 }
