--- apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/entities/Condition.java
@@ -1,29 +1,0 @@
-package gov.cdc.nbs.questionbank.entities;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "condition", catalog = "question_bank")
-public class Condition {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "description", length = 300)
-    private String description;
-
-}
