+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/questionnaire/Questionnaire.java
@@ -1,8 +1,10 @@
-package gov.cdc.nbs.questionbank.entities;
+package gov.cdc.nbs.questionbank.questionnaire;
 
 import java.util.List;
 import javax.persistence.CascadeType;
+import javax.persistence.CollectionTable;
 import javax.persistence.Column;
+import javax.persistence.ElementCollection;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -12,6 +14,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
+import gov.cdc.nbs.questionbank.entities.QuestionGroup;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -25,12 +28,11 @@ public class Questionnaire {
     @Column(name = "id", nullable = false)
     private Long id;
 
-    @OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinTable(
+    @ElementCollection
+    @CollectionTable(
             name = "questionnaire_conditions",
-            joinColumns = @JoinColumn(name = "questionnaire_id"),
-            inverseJoinColumns = @JoinColumn(name = "condition_id"))
-    private List<Condition> conditions;
+            joinColumns = @JoinColumn(name = "condition_cd"))
+    private List<String> conditionCodes;
 
     @Column(name = "questionnaire_type", length = 100, nullable = false)
     private String questionnaireType;
