+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/questionnaire/QuestionnaireContext.java
@@ -1,5 +1,5 @@
 package gov.cdc.nbs.questionbank.questionnaire;
 
-public record QuestionnaireContext(long conditionId, String questionnaireType) {
+public record QuestionnaireContext(String conditionCd, String questionnaireType) {
 
 }
