+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/questionnaire/QuestionnaireResolver.java
@@ -3,7 +3,6 @@ package gov.cdc.nbs.questionbank.questionnaire;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
-import gov.cdc.nbs.questionbank.entities.Questionnaire;
 
 @Controller
 public class QuestionnaireResolver {
