+++ apps/question-bank/src/test/java/gov/cdc/nbs/questionbank/support/QuestionnaireMother.java
@@ -1,17 +1,16 @@
 package gov.cdc.nbs.questionbank.support;
 
 import java.util.List;
-import gov.cdc.nbs.questionbank.entities.Condition;
 import gov.cdc.nbs.questionbank.entities.QuestionGroup;
-import gov.cdc.nbs.questionbank.entities.Questionnaire;
+import gov.cdc.nbs.questionbank.questionnaire.Questionnaire;
 
 public class QuestionnaireMother {
 
     public static Questionnaire questionnaire(
             String type, List<QuestionGroup> questionGroups,
-            List<Condition> conditions) {
+            List<String> conditions) {
         var q = new Questionnaire();
-        q.setConditions(conditions);
+        q.setConditionCodes(conditions);
         q.setQuestionnaireType(type);
         q.setQuestionGroups(questionGroups);
         return q;
